+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Release date: ``2019-10-29``
 - `NXPY-128 <https://jira.nuxeo.com/browse/NXPY-128>`__: Make ``Batch`` upload index public
 - `NXPY-135 <https://jira.nuxeo.com/browse/NXPY-135>`__: Expand the documentation on how to attach multiple blobs to a given document
 - `NXPY-136 <https://jira.nuxeo.com/browse/NXPY-136>`__: Allow additionnal parameters to ``Directories.get()``
+- `NXPY-137 <https://jira.nuxeo.com/browse/NXPY-137>`__: Fix failing test about converters
 
 Technical changes
 -----------------